Arrumaçao geral Makefile e detalhes

+++ ft_strncmp.c
@@ -6,13 +6,15 @@
 /*   By: julrodri <julrodri@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/30 17:39:52 by julrodri          #+#    #+#             */
-/*   Updated: 2021/09/01 23:56:56 by julrodri         ###   ########.fr       */
+/*   Updated: 2021/09/02 23:11:32 by julrodri         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "libft.h"
+
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 	int	r;
 
 	if (n == 0)
