+++ ft_strnstr.c
@@ -6,9 +6,11 @@
 /*   By: julrodri <julrodri@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/02 00:04:56 by julrodri          #+#    #+#             */
-/*   Updated: 2021/09/02 00:06:20 by julrodri         ###   ########.fr       */
+/*   Updated: 2021/09/02 23:17:22 by julrodri         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+
+#include "libft.h"
 
 static void	check_whole_str(
 const char *str, const char *to_find, int *i, int *mem)
@@ -41,7 +43,7 @@ char	*ft_strnstr(const char *haystack, const char *needle, size_t n)
 	r = (char *) &haystack[i];
 	if (needle[0] == '\0')
 		return (r);
-	while (haystack[i] != '\0' && i >= 0 && i < n)
+	while (haystack[i] != '\0' && i >= 0 && i < (int) n)
 	{
 		check_whole_str(haystack, needle, &i, &mem);
 		i++;
